Added scroll view

+++ Vx.iOS/BareUIHelper.cs
@@ -12,7 +12,7 @@ namespace Vx.iOS
     {
         public const int STATUS_AND_NAV_BAR_HEIGHT = 44;
 
-        public static UIView ConfigureForVerticalScrolling(this UIView view, UIScrollView parentScrollView, int left = 0, int top = 0, int right = 0, int bottom = 0)
+        public static UIView ConfigureForVerticalScrolling(this UIView view, UIScrollView parentScrollView, float left = 0, float top = 0, float right = 0, float bottom = 0)
         {
             // https://developer.apple.com/library/content/documentation/UserExperience/Conceptual/AutolayoutPG/WorkingwithScrollViews.html#//apple_ref/doc/uid/TP40010853-CH24-SW1
